Decouple markdown validation
