+++ src/docfx/SubCommands/DocumentBuilderWrapper.cs
@@ -26,7 +26,6 @@ namespace Microsoft.DocAsCode.SubCommands
     using Microsoft.DocAsCode.Common;
     using Microsoft.DocAsCode.Exceptions;
     using Microsoft.DocAsCode.Plugins;
-    using Microsoft.DocAsCode.MarkdigEngine;
 
     [Serializable]
     internal sealed class DocumentBuilderWrapper
@@ -156,7 +155,6 @@ namespace Microsoft.DocAsCode.SubCommands
                 typeof(TocDocumentProcessor).Assembly,
                 typeof(SchemaDrivenDocumentProcessor).Assembly,
                 typeof(UniversalReferenceDocumentProcessor).Assembly,
-                typeof(MarkdigServiceProvider).Assembly
             };
             foreach (var assem in defaultPluggedAssemblies)
             {
