+++ test/Microsoft.DocAsCode.MarkdigEngine.Tests/TestUtility/TestUtility.cs
@@ -9,6 +9,7 @@ namespace Microsoft.DocAsCode.MarkdigEngine.Tests
 
     using MarkdigEngine.Extensions;
 
+    using Microsoft.DocAsCode.Build.Engine;
     using Microsoft.DocAsCode.Plugins;
     using Xunit;
 
@@ -32,7 +33,7 @@ namespace Microsoft.DocAsCode.MarkdigEngine.Tests
             {
                 BasePath = "."
             };
-            var service = new MarkdigMarkdownService(parameter);
+            var service = new MarkdigMarkdownServiceCreator().CreateMarkdigMarkdownService(parameter);
 
             return service.Markup(content, filePath ?? string.Empty);
         }
@@ -47,7 +48,7 @@ namespace Microsoft.DocAsCode.MarkdigEngine.Tests
                     { LineNumberExtension.EnableSourceInfo, false }
                 }
             };
-            var service = new MarkdigMarkdownService(parameter);
+            var service = new MarkdigMarkdownServiceCreator().CreateMarkdigMarkdownService(parameter);
 
             return service.Markup(content, filePath ?? string.Empty);
         }
@@ -74,7 +75,7 @@ namespace Microsoft.DocAsCode.MarkdigEngine.Tests
                 }
             };
 
-            return new MarkdigMarkdownService(parameter);
+            return new MarkdigMarkdownServiceCreator().CreateMarkdigMarkdownService(parameter);
         }
     }
 }
